Makes PropertyPad platform agnostic with Gtk and Cocoa implementations
