+++ main/src/addins/MonoDevelop.DesignerSupport/MonoDevelop.DesignerSupport.addin.xml
@@ -47,7 +47,7 @@
 			defaultStatus="AutoHide"
 			/>
 		<Pad id = "MonoDevelop.DesignerSupport.PropertyPad"
-			class = "MonoDevelop.DesignerSupport.MacPropertyPad"
+			class = "MonoDevelop.DesignerSupport.PropertyPad"
 			_label="Properties"
 			icon = "md-properties-pad"
 			group = "Designer"
