+++ main/src/addins/MonoDevelop.DesignerSupport/MonoDevelop.DesignerSupport/IPropertyPad.cs
@@ -1,0 +1,16 @@
+﻿
+using System;
+
+namespace MonoDevelop.DesignerSupport
+{
+	public interface IPropertyPad
+	{
+		bool IsPropertyGridEditing { get; }
+
+		event EventHandler PropertyGridChanged;
+
+		void SetCurrentObject (object lastComponent, object [] propertyProviders);
+		void BlankPad ();
+		void PopulateGrid (bool saveEditSession);
+	}
+}
